worked on the UI for buttons and scenes
worked on the UI for buttons and scenes
the GameOver panel and Endgame scene are reading to be invoke on the game. Currently will require interface with player health to initiate GAME OVER panel and player to collide to portal to SetActive Endgame panel. Code are currently comment and ready to add the line of code on playerzombie interface is complete.